feat: move SourceSystemPrefix to common package
