+++ src/main/java/egovframework/bat/insa/processor/EmployeeInfoProcessor.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 
 import egovframework.bat.insa.domain.EmployeeInfo;
 import egovframework.bat.insa.util.EsntlIdGenerator;
-import egovframework.bat.insa.domain.SourceSystemPrefix;
+import egovframework.bat.insa.common.SourceSystemPrefix;
 import lombok.RequiredArgsConstructor;
 
 /**
